+++ CHOLMOD/MATLAB/chol2.c
@@ -7,7 +7,7 @@
  * The CHOLMOD/MATLAB Module is licensed under Version 2.0 of the GNU
  * General Public License.  See gpl.txt for a text of the license.
  * CHOLMOD is also available under other licenses; contact authors for details.
- * http://www.cise.ufl.edu/research/sparse
+ * http://www.suitesparse.com
  * MATLAB(tm) is a Trademark of The MathWorks, Inc.
  * -------------------------------------------------------------------------- */
 
@@ -42,7 +42,7 @@ void mexFunction
     cholmod_sparse Amatrix, *A, *Lsparse, *R ;
     cholmod_factor *L ;
     cholmod_common Common, *cm ;
-    Int n, minor ;
+    Long n, minor ;
 
     /* ---------------------------------------------------------------------- */
     /* start CHOLMOD and set parameters */
