+++ CHOLMOD/MATLAB/nesdis.c
@@ -7,7 +7,7 @@
  * The CHOLMOD/MATLAB Module is licensed under Version 2.0 of the GNU
  * General Public License.  See gpl.txt for a text of the license.
  * CHOLMOD is also available under other licenses; contact authors for details.
- * http://www.cise.ufl.edu/research/sparse
+ * http://www.suitesparse.com
  * MATLAB(tm) is a Trademark of The MathWorks, Inc.
  * METIS (Copyright 1998, G. Karypis) is not distributed with CHOLMOD.
  * -------------------------------------------------------------------------- */
@@ -63,10 +63,10 @@ void mexFunction
 {
 #ifndef NPARTITION
     double dummy = 0 ;
-    Int *Perm, *Cmember, *CParent ;
+    Long *Perm, *Cmember, *CParent ;
     cholmod_sparse *A, Amatrix, *C, *S ;
     cholmod_common Common, *cm ;
-    Int n, transpose, c, ncomp ;
+    Long n, transpose, c, ncomp ;
     char buf [LEN] ;
 
     /* ---------------------------------------------------------------------- */
@@ -164,9 +164,9 @@ void mexFunction
     /* get workspace */
     /* ---------------------------------------------------------------------- */
 
-    CParent = cholmod_l_malloc (n, sizeof (Int), cm) ;
-    Cmember = cholmod_l_malloc (n, sizeof (Int), cm) ;
-    Perm = cholmod_l_malloc (n, sizeof (Int), cm) ;
+    CParent = cholmod_l_malloc (n, sizeof (Long), cm) ;
+    Cmember = cholmod_l_malloc (n, sizeof (Long), cm) ;
+    Perm = cholmod_l_malloc (n, sizeof (Long), cm) ;
 
     /* ---------------------------------------------------------------------- */
     /* order the matrix with CHOLMOD's nested dissection */
@@ -197,9 +197,9 @@ void mexFunction
     /* free workspace */
     /* ---------------------------------------------------------------------- */
 
-    cholmod_l_free (n, sizeof (Int), Perm, cm) ;
-    cholmod_l_free (n, sizeof (Int), CParent, cm) ;
-    cholmod_l_free (n, sizeof (Int), Cmember, cm) ;
+    cholmod_l_free (n, sizeof (Long), Perm, cm) ;
+    cholmod_l_free (n, sizeof (Long), CParent, cm) ;
+    cholmod_l_free (n, sizeof (Long), Cmember, cm) ;
     cholmod_l_free_sparse (&C, cm) ;
     cholmod_l_free_sparse (&S, cm) ;
     cholmod_l_finish (cm) ;
