+++ CHOLMOD/MATLAB/septree.c
@@ -7,7 +7,7 @@
  * The CHOLMOD/MATLAB Module is licensed under Version 2.0 of the GNU
  * General Public License.  See gpl.txt for a text of the license.
  * CHOLMOD is also available under other licenses; contact authors for details.
- * http://www.cise.ufl.edu/research/sparse
+ * http://www.suitesparse.com
  * MATLAB(tm) is a Trademark of The MathWorks, Inc.
  * METIS (Copyright 1998, G. Karypis) is not distributed with CHOLMOD.
  * -------------------------------------------------------------------------- */
@@ -46,10 +46,10 @@ void mexFunction
 {
 #ifndef NPARTITION
     double *p ;
-    Int *Cmember, *CParent ;
+    Long *Cmember, *CParent ;
     cholmod_common Common, *cm ;
     double nd_oksep ;
-    Int nd_small, nc, n, c, j, nc_new ;
+    Long nd_small, nc, n, c, j, nc_new ;
 
     /* ---------------------------------------------------------------------- */
     /* start CHOLMOD and set defaults */
@@ -79,8 +79,8 @@ void mexFunction
 	mexErrMsgTxt ("invalid inputs") ;
     }
 
-    CParent = cholmod_l_malloc (nc, sizeof (Int), cm) ;
-    Cmember = cholmod_l_malloc (n, sizeof (Int), cm) ;
+    CParent = cholmod_l_malloc (nc, sizeof (Long), cm) ;
+    Cmember = cholmod_l_malloc (n, sizeof (Long), cm) ;
 
     p = mxGetPr (pargin [0]) ;
     for (c = 0 ; c < nc ; c++)
@@ -128,8 +128,8 @@ void mexFunction
     /* free workspace */
     /* ---------------------------------------------------------------------- */
 
-    cholmod_l_free (nc, sizeof (Int), CParent, cm) ;
-    cholmod_l_free (n, sizeof (Int), Cmember, cm) ;
+    cholmod_l_free (nc, sizeof (Long), CParent, cm) ;
+    cholmod_l_free (n, sizeof (Long), Cmember, cm) ;
     cholmod_l_finish (cm) ;
     cholmod_l_print_common (" ", cm) ;
     /*
