+++ CHOLMOD/MATLAB/symbfact2.c
@@ -7,7 +7,7 @@
  * The CHOLMOD/MATLAB Module is licensed under Version 2.0 of the GNU
  * General Public License.  See gpl.txt for a text of the license.
  * CHOLMOD is also available under other licenses; contact authors for details.
- * http://www.cise.ufl.edu/research/sparse
+ * http://www.suitesparse.com
  * MATLAB(tm) is a Trademark of The MathWorks, Inc.
  * -------------------------------------------------------------------------- */
 
@@ -51,10 +51,10 @@ void mexFunction
 {
     double dummy = 0 ;
     double *Lx, *px ;
-    Int *Parent, *Post, *ColCount, *First, *Level, *Rp, *Ri, *Lp, *Li, *W ;
+    Long *Parent, *Post, *ColCount, *First, *Level, *Rp, *Ri, *Lp, *Li, *W ;
     cholmod_sparse *A, Amatrix, *F, *Aup, *Alo, *R, *A1, *A2, *L, *S ;
     cholmod_common Common, *cm ;
-    Int n, i, coletree, j, lnz, p, k, height, c ;
+    Long n, i, coletree, j, lnz, p, k, height, c ;
     char buf [LEN] ;
 
     /* ---------------------------------------------------------------------- */
@@ -134,11 +134,11 @@ void mexFunction
     /* compute the etree, its postorder, and the row/column counts */
     /* ---------------------------------------------------------------------- */
 
-    Parent = cholmod_l_malloc (n, sizeof (Int), cm) ;
-    Post = cholmod_l_malloc (n, sizeof (Int), cm) ;
-    ColCount = cholmod_l_malloc (n, sizeof (Int), cm) ;
-    First = cholmod_l_malloc (n, sizeof (Int), cm) ;
-    Level = cholmod_l_malloc (n, sizeof (Int), cm) ;
+    Parent = cholmod_l_malloc (n, sizeof (Long), cm) ;
+    Post = cholmod_l_malloc (n, sizeof (Long), cm) ;
+    ColCount = cholmod_l_malloc (n, sizeof (Long), cm) ;
+    First = cholmod_l_malloc (n, sizeof (Long), cm) ;
+    Level = cholmod_l_malloc (n, sizeof (Long), cm) ;
 
     /* F = A' */
     F = cholmod_l_transpose (A, 0, cm) ;
@@ -321,11 +321,11 @@ void mexFunction
     /* free workspace */
     /* ---------------------------------------------------------------------- */
 
-    cholmod_l_free (n, sizeof (Int), Parent, cm) ;
-    cholmod_l_free (n, sizeof (Int), Post, cm) ;
-    cholmod_l_free (n, sizeof (Int), ColCount, cm) ;
-    cholmod_l_free (n, sizeof (Int), First, cm) ;
-    cholmod_l_free (n, sizeof (Int), Level, cm) ;
+    cholmod_l_free (n, sizeof (Long), Parent, cm) ;
+    cholmod_l_free (n, sizeof (Long), Post, cm) ;
+    cholmod_l_free (n, sizeof (Long), ColCount, cm) ;
+    cholmod_l_free (n, sizeof (Long), First, cm) ;
+    cholmod_l_free (n, sizeof (Long), Level, cm) ;
     cholmod_l_free_sparse (&F, cm) ;
     cholmod_l_free_sparse (&S, cm) ;
     cholmod_l_finish (cm) ;
