+++ SPQR/MATLAB/spqr_qmult.cpp
@@ -32,9 +32,9 @@ void mexFunction
 )
 {
     mxArray *Hmatlab, *Tau, *P ;
-    Int *HPinv, *Yp, *Yi ;
+    Long *HPinv, *Yp, *Yi ;
     double *Hx, *Xx, *Tx, *Px, dummy ;
-    Int m, n, k, nh, nb, p, i, method, mh, gotP, X_is_sparse, is_complex, hnz,
+    Long m, n, k, nh, nb, p, i, method, mh, gotP, X_is_sparse, is_complex, hnz,
         tnz, xnz, inuse, count ;
     cholmod_sparse *Ysparse, *H, Hmatrix, *Xsparse, Xsmatrix ;
     cholmod_dense *Ydense, *Xdense, Xdmatrix, *HTau, HTau_matrix ;
@@ -142,7 +142,7 @@ void mexFunction
     }
     else
     {
-        method = (Int) mxGetScalar (pargin [2]) ;
+        method = (Long) mxGetScalar (pargin [2]) ;
         if (method < 0 || method > 3)
         {
             mexErrMsgIdAndTxt ("QR:invalidInput", "invalid method") ;
@@ -215,11 +215,11 @@ void mexFunction
             mexErrMsgIdAndTxt ("QR:invalidInput",
                 "P must be a vector of length equal to # rows of H") ;
         }
-        HPinv = (Int *) cholmod_l_malloc (mh, sizeof (Int), cc) ;
+        HPinv = (Long *) cholmod_l_malloc (mh, sizeof (Long), cc) ;
         Px = mxGetPr (P) ;
         for (i = 0 ; i < mh ; i++)
         {
-            HPinv [i] = (Int) (Px [i] - 1) ;
+            HPinv [i] = (Long) (Px [i] - 1) ;
             if (HPinv [i] < 0 || HPinv [i] >= mh)
             {
                 mexErrMsgIdAndTxt ("QR:invalidInput", "invalid permutation") ;
@@ -266,7 +266,7 @@ void mexFunction
     // free workspace
     // -------------------------------------------------------------------------
 
-    cholmod_l_free (mh, sizeof (Int), HPinv, cc) ;
+    cholmod_l_free (mh, sizeof (Long), HPinv, cc) ;
 
     if (is_complex)
     {
